feat: migrate validator to multiplatform
