+++ features/auth/presentation/build.gradle.kts
@@ -62,7 +62,7 @@ dependencies {
     implementation(projects.foundation.network)
     implementation(projects.foundation.navigation)
     implementation(projects.foundation.designSystem)
-    implementation(projects.foundation.validator.impl)
+    implementation(projects.foundation.validator)
     implementation(projects.foundation.ui)
     implementation(projects.foundation.exception)
     implementation(projects.foundation.auth)
