+++ features/beneficiary/presentation/src/test/kotlin/io/github/alaksion/invoicer/features/beneficiary/presentation/screen/create/CreateBeneficiaryScreenModelTest.kt
@@ -1,7 +1,7 @@
 package io.github.alaksion.invoicer.features.beneficiary.presentation.screen.create
 
 import foundation.network.RequestError
-import foundation.watchers.RefreshBeneficiaryPublisher
+import io.github.alaksion.invoicer.foundation.watchers.RefreshBeneficiaryPublisher
 import io.github.alaksion.invoicer.features.beneficiary.services.domain.repository.BeneficiaryRepository
 import io.mockk.Runs
 import io.mockk.clearAllMocks
