+++ features/intermediary/presentation/src/main/kotlin/io/github/alaksion/invoicer/features/intermediary/presentation/screen/create/CreateIntermediaryScreenModel.kt
@@ -4,7 +4,7 @@ import cafe.adriel.voyager.core.model.ScreenModel
 import cafe.adriel.voyager.core.model.screenModelScope
 import foundation.network.request.handle
 import foundation.network.request.launchRequest
-import foundation.watchers.RefreshIntermediaryPublisher
+import io.github.alaksion.invoicer.foundation.watchers.RefreshIntermediaryPublisher
 import io.github.alaksion.invoicer.features.intermediary.services.domain.repository.IntermediaryRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Job
