+++ features/intermediary/presentation/src/test/kotlin/io/github/alaksion/invoicer/features/intermediary/presentation/screen/create/CreateIntermediaryScreenModelTest.kt
@@ -1,7 +1,7 @@
 package io.github.alaksion.invoicer.features.intermediary.presentation.screen.create
 
 import foundation.network.RequestError
-import foundation.watchers.RefreshIntermediaryPublisher
+import io.github.alaksion.invoicer.foundation.watchers.RefreshIntermediaryPublisher
 import io.github.alaksion.invoicer.features.intermediary.services.domain.repository.IntermediaryRepository
 import io.mockk.Runs
 import io.mockk.clearAllMocks
@@ -12,7 +12,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.test.setMain
