+++ features/invoice/src/main/kotlin/io/github/alaksion/invoicer/features/invoice/presentation/screens/create/steps/pickbeneficiary/PickBeneficiaryScreen.kt
@@ -27,7 +27,7 @@ import cafe.adriel.voyager.koin.koinScreenModel
 import cafe.adriel.voyager.navigator.LocalNavigator
 import foundation.navigation.InvoicerScreen
 import io.github.alaksion.invoicer.foundation.ui.events.EventEffect
-import foundation.watchers.RefreshBeneficiaryPublisher
+import io.github.alaksion.invoicer.foundation.watchers.RefreshBeneficiaryPublisher
 import io.github.alaksion.invoicer.features.invoice.presentation.screens.create.components.SelectableItem
 import io.github.alaksion.invoicer.features.invoice.presentation.screens.create.steps.pickintermediary.PickIntermediaryScreen
 import io.github.alaksion.invoicer.foundation.designSystem.components.LoadingState
