+++ foundation/auth/src/main/kotlin/io/github/alaksion/invoicer/foundation/auth/domain/service/SignInCommandManager.kt
@@ -1,7 +1,7 @@
 package io.github.alaksion.invoicer.foundation.auth.domain.service
 
-import foundation.watchers.AuthEvent
-import foundation.watchers.AuthEventBus
+import io.github.alaksion.invoicer.foundation.watchers.AuthEvent
+import io.github.alaksion.invoicer.foundation.watchers.AuthEventBus
 import io.github.alaksion.invoicer.foundation.auth.domain.repository.AuthRepository
 import io.github.alaksion.invoicer.foundation.auth.domain.repository.AuthTokenRepository
 import io.github.alaksion.invoicer.foundation.session.Session
