--- foundation/validator/impl/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    id("invoicer.library")
-}
-
-android {
-    namespace = "foundation.validator.impl"
-}
-
-
-dependencies {
-    implementation(platform(libs.koin.bom))
-    implementation(libs.koin.core)
-}
