--- foundation/validator/impl/src/main/kotlin/foundation/validator/impl/di/ValidatorDiModule.kt
@@ -1,9 +1,0 @@
-package foundation.validator.impl.di
-
-import foundation.validator.impl.UuidValidator
-import foundation.validator.impl.UuidValidatorImpl
-import org.koin.dsl.module
-
-val validatorDiModule = module {
-    factory<UuidValidator> { UuidValidatorImpl() }
-}
