--- foundation/validator/test/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    id("invoicer.library")
-}
-
-android {
-    namespace = "foundation.validator.test"
-}
-
-dependencies {
-    implementation(platform(libs.koin.bom))
-    implementation(libs.koin.core)
-}
