+++ foundation/watchers/src/commonMain/kotlin/io/github/alaksion/invoicer/foundation/watchers/NewInvoicePublisher.kt
@@ -1,6 +1,6 @@
-package foundation.watchers
+package io.github.alaksion.invoicer.foundation.watchers
 
 import io.github.alaksion.invoicer.foundation.ui.events.EventAware
 import io.github.alaksion.invoicer.foundation.ui.events.EventPublisher
 
-class RefreshIntermediaryPublisher : EventAware<Unit> by EventPublisher()
+class NewInvoicePublisher : EventAware<Unit> by EventPublisher()
