+++ settings.gradle.kts
@@ -44,8 +44,7 @@ include(":features:auth:presentation")
 include(":foundation:navigation")
 include(":foundation:network")
 include(":foundation:exception")
-include(":foundation:validator:impl")
-include(":foundation:validator:test")
+include(":foundation:validator")
 include(":foundation:storage:impl")
 include(":foundation:storage:test")
 include(":foundation:session")
